Refactor: Move extractCurrentStep and all associated functions to BlazeWebViewViewModel.
